fix: parent field of a backing image returned to the instance manager should be empty
The parent field of a backing image returned to the instance manager should be empty.

Longhorn 10526
